Extract `ProjectStateImpl` into own `DefaultProjectState` top-level class
